Fixed turning on/off airplane mode, and sending saving this mode status with each recording.  Seems to detect sim card OK - just do this on MainActivity resume as you have to turn airplane mode off to detect sim which meant I was doing this on every recording.
If there is no connection to server, files are no longer deleted, but sent next time there is a connection.

+++ app/src/main/java/nz/org/cacophonoy/cacophonometerlite/UploadFiles.java
@@ -5,11 +5,15 @@ import android.content.Context;
 import android.media.AudioManager;
 import android.media.ToneGenerator;
 import android.util.Log;
+import android.widget.Toast;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.File;
+
+import static nz.org.cacophonoy.cacophonometerlite.Server.disableDataConnection;
+import static nz.org.cacophonoy.cacophonometerlite.Server.enableDataConnection;
 
 class UploadFiles implements Runnable { // http://stackoverflow.com/questions/15666260/urlconnection-android-4-2-doesnt-work
     private static final String LOG_TAG = UploadFiles.class.getName();
@@ -24,10 +28,10 @@ class UploadFiles implements Runnable { // http://stackoverflow.com/questions/15
         this.context = context;
     }
 
-
-    private void sendFile(File aFile) {
-        System.out.println("Going to upload file " + aFile.getName());
+    private boolean sendFile(File aFile) {
+  //  private void sendFile(File aFile) {
         Log.i(LOG_TAG, "sendFile: Start");
+        Prefs prefs = new Prefs(context);
 
         // Get metadata and put into JSON.
         JSONObject audioRecording = new JSONObject();
@@ -45,20 +49,17 @@ class UploadFiles implements Runnable { // http://stackoverflow.com/questions/15
         String second = fileNameParts[5];
         String relativeTo = fileNameParts[6];
         String relativeToOffset = fileNameParts[7];
+        String airplaneMode = fileNameParts[8];
+        boolean airplaneModeOn = false;
+        if (airplaneMode.equalsIgnoreCase("airplaneModeOn")){
+            airplaneModeOn =  true;
+        }
 
         String localFilePath = "/data/data/com.thecacophonytrust.cacophonometer/app_cacophony/recordings/" + fileName;
         String recordingDateTime = year + "-" + month + "-" + day + " " + hour + ":" + minute + ":" + second;
         String recordingTime = hour + ":" + minute + ":" + second;
-
-        String  batteryStatus = Util.getBatteryStatus(context);
-
-      //  float batteryLevel = Util.getBatteryLevel(context);
-
-        boolean airplaneModeOn = Util.isAirplaneModeOn(context);
-
-
-        Prefs prefs = new Prefs(context);
-
+        String batteryStatus = Util.getBatteryStatus(context);
+       // boolean airplaneModeOn = Util.isAirplaneModeOn(context);
 
         try {
             audioRecording.put("location", "Lat: " + prefs.getLatitude() + ", Lon: " + prefs.getLongitude());
@@ -66,52 +67,49 @@ class UploadFiles implements Runnable { // http://stackoverflow.com/questions/15
             audioRecording.put("localFilePath", localFilePath);
             audioRecording.put("recordingDateTime", recordingDateTime);
             audioRecording.put("recordingTime", recordingTime);
-
             audioRecording.put("batteryCharging", batteryStatus);
-          //  audioRecording.put("batteryLevel", batteryLevel);
             audioRecording.put("batteryLevel", prefs.getBatteryLevel());
             audioRecording.put("airplaneModeOn", airplaneModeOn);
-            if (relativeTo.equalsIgnoreCase("relativeToDawn")){
+            if (relativeTo.equalsIgnoreCase("relativeToDawn")) {
                 audioRecording.put("relativeToDawn", relativeToOffset);
             }
-            if (relativeTo.equalsIgnoreCase("relativeToDusk")){
+            if (relativeTo.equalsIgnoreCase("relativeToDusk")) {
                 audioRecording.put("relativeToDusk", relativeToOffset);
             }
-//            audioRecording.put("relativeToDawn", );
-//            audioRecording.put("relativeToDusk", );
-
-            // hack for now to see battery status
-        //    audioRecording.put("location", "Battery status: " + batteryStatus + ", Battery Level: " + batteryLevel + ", airplaneModeOn: " + airplaneModeOn);
 
 
         } catch (JSONException e) {
             e.printStackTrace();
         }
 
-        Server.uploadAudioRecording(aFile, audioRecording, context);
+//        Server.uploadAudioRecording(aFile, audioRecording, context);
+        return Server.uploadAudioRecording(aFile, audioRecording, context);
     }
 
     @Override
     public void run() {
 
-        ToneGenerator toneG = new ToneGenerator(AudioManager.STREAM_ALARM, ToneGenerator.MAX_VOLUME);
-        toneG.startTone(ToneGenerator.TONE_CDMA_DIAL_TONE_LITE, 5000);
-
-
-
-            File recordingsFolder = Util.getRecordingsFolder();
-            File recordingFiles[] = recordingsFolder.listFiles();
-            if (recordingFiles != null) {
-                for (File aFile : recordingFiles) {
-                    System.out.println("File name is " + aFile.getName());
-                    sendFile(aFile);
-                    if (!aFile.delete())
-                        Log.w(LOG_TAG, "Deleting audio file failed");
-                }
-            }
+        File recordingsFolder = Util.getRecordingsFolder();
+        File recordingFiles[] = recordingsFolder.listFiles();
+        if (recordingFiles != null) {
+            enableDataConnection(context);
+          //  Toast.makeText(context, "Number of Files to upload is " + recordingFiles.length,  Toast.LENGTH_LONG).show();
+            for (File aFile : recordingFiles) {
+//                sendFile(aFile);
+//                if (!aFile.delete())
+//                    Log.w(LOG_TAG, "Deleting audio file failed");
+               if (sendFile(aFile)){
+                   if (!aFile.delete()){
+                       Log.w(LOG_TAG, "Deleting audio file failed");
                    }
+               }else{
+                   Log.w(LOG_TAG, "Failed to upload file to server");
+               }
 
-
+            }
+        }
+        disableDataConnection(context);
+    }
 
 
 }
